# configuration warning rework