--- a/core/extension/gdextension_interface.h
+++ b/core/extension/gdextension_interface.h
@@ -256,12 +256,21 @@ typedef struct {
 	GDExtensionVariantPtr *default_arguments;
 } GDExtensionMethodInfo;
 
+typedef struct {
+	GDExtensionStringPtr message;
+	GDExtensionStringNamePtr code;
+	GDExtensionStringNamePtr property_name;
+	int32_t severity;
+} GDExtensionConfigurationInfo;
+
 typedef const GDExtensionPropertyInfo *(*GDExtensionClassGetPropertyList)(GDExtensionClassInstancePtr p_instance, uint32_t *r_count);
 typedef void (*GDExtensionClassFreePropertyList)(GDExtensionClassInstancePtr p_instance, const GDExtensionPropertyInfo *p_list);
 typedef void (*GDExtensionClassFreePropertyList2)(GDExtensionClassInstancePtr p_instance, const GDExtensionPropertyInfo *p_list, uint32_t p_count);
 typedef GDExtensionBool (*GDExtensionClassPropertyCanRevert)(GDExtensionClassInstancePtr p_instance, GDExtensionConstStringNamePtr p_name);
 typedef GDExtensionBool (*GDExtensionClassPropertyGetRevert)(GDExtensionClassInstancePtr p_instance, GDExtensionConstStringNamePtr p_name, GDExtensionVariantPtr r_ret);
 typedef GDExtensionBool (*GDExtensionClassValidateProperty)(GDExtensionClassInstancePtr p_instance, GDExtensionPropertyInfo *p_property);
+typedef const GDExtensionConfigurationInfo *(*GDExtensionClassGetConfigurationInfo)(GDExtensionClassInstancePtr p_instance, uint32_t *r_count);
+typedef void (*GDExtensionClassFreeConfigurationInfo)(GDExtensionClassInstancePtr p_instance, const GDExtensionConfigurationInfo *p_list, uint32_t p_count);
 typedef void (*GDExtensionClassNotification)(GDExtensionClassInstancePtr p_instance, int32_t p_what); // Deprecated. Use GDExtensionClassNotification2 instead.
 typedef void (*GDExtensionClassNotification2)(GDExtensionClassInstancePtr p_instance, int32_t p_what, GDExtensionBool p_reversed);
 typedef void (*GDExtensionClassToString)(GDExtensionClassInstancePtr p_instance, GDExtensionBool *r_is_valid, GDExtensionStringPtr p_out);
@@ -379,6 +388,8 @@ typedef struct {
 	GDExtensionClassPropertyGetRevert property_get_revert_func;
 	GDExtensionClassValidateProperty validate_property_func;
 	GDExtensionClassNotification2 notification_func;
+	GDExtensionClassGetConfigurationInfo get_configuration_info_func;
+	GDExtensionClassFreeConfigurationInfo free_configuration_info_func;
 	GDExtensionClassToString to_string_func;
 	GDExtensionClassReference reference_func;
 	GDExtensionClassUnreference unreference_func;
@@ -585,6 +596,8 @@ typedef GDExtensionInt (*GDExtensionScriptInstanceGetMethodArgumentCount)(GDExte
 typedef void (*GDExtensionScriptInstanceCall)(GDExtensionScriptInstanceDataPtr p_self, GDExtensionConstStringNamePtr p_method, const GDExtensionConstVariantPtr *p_args, GDExtensionInt p_argument_count, GDExtensionVariantPtr r_return, GDExtensionCallError *r_error);
 typedef void (*GDExtensionScriptInstanceNotification)(GDExtensionScriptInstanceDataPtr p_instance, int32_t p_what); // Deprecated. Use GDExtensionScriptInstanceNotification2 instead.
 typedef void (*GDExtensionScriptInstanceNotification2)(GDExtensionScriptInstanceDataPtr p_instance, int32_t p_what, GDExtensionBool p_reversed);
+typedef const GDExtensionConfigurationInfo *(*GDExtensionScriptInstanceGetConfigurationInfo)(GDExtensionScriptInstanceDataPtr p_instance, uint32_t *r_count);
+typedef void (*GDExtensionScriptInstanceFreeConfigurationInfo)(GDExtensionScriptInstanceDataPtr p_instance, const GDExtensionConfigurationInfo *p_list, uint32_t p_count);
 typedef void (*GDExtensionScriptInstanceToString)(GDExtensionScriptInstanceDataPtr p_instance, GDExtensionBool *r_is_valid, GDExtensionStringPtr r_out);
 
 typedef void (*GDExtensionScriptInstanceRefCountIncremented)(GDExtensionScriptInstanceDataPtr p_instance);
@@ -705,6 +718,8 @@ typedef struct {
 
 	GDExtensionScriptInstanceCall call_func;
 	GDExtensionScriptInstanceNotification2 notification_func;
+	GDExtensionScriptInstanceGetConfigurationInfo get_configuration_info_func;
+	GDExtensionScriptInstanceFreeConfigurationInfo free_configuration_info_func;
 
 	GDExtensionScriptInstanceToString to_string_func;