--- a/core/object/script_instance.h
+++ b/core/object/script_instance.h
@@ -69,6 +69,7 @@ public:
 
 	virtual Variant call_const(const StringName &p_method, const Variant **p_args, int p_argcount, Callable::CallError &r_error); // implement if language supports const functions
 	virtual void notification(int p_notification, bool p_reversed = false) = 0;
+	virtual void get_configuration_info(List<ConfigurationInfo> *p_infos) const = 0;
 	virtual String to_string(bool *r_valid) {
 		if (r_valid) {
 			*r_valid = false;