--- a/core/object/script_language.h
+++ b/core/object/script_language.h
@@ -478,6 +478,7 @@ public:
 
 	virtual Variant callp(const StringName &p_method, const Variant **p_args, int p_argcount, Callable::CallError &r_error) override;
 	virtual void notification(int p_notification, bool p_reversed = false) override {}
+	virtual void get_configuration_info(List<ConfigurationInfo> *p_infos) const override {}
 
 	virtual Ref<Script> get_script() const override { return script; }