--- a/editor/editor_inspector.h
+++ b/editor/editor_inspector.h
@@ -97,7 +97,6 @@ private:
 	bool checkable = false;
 	bool checked = false;
 	bool draw_warning = false;
-	bool draw_prop_warning = false;
 	bool keying = false;
 	bool deletable = false;