… core/object/script_language_extension.h
… doc/classes/Node.xml
… doc/classes/Object.xml
… doc/classes/SceneTree.xml
… editor/editor_configuration_info.cpp
… editor/editor_configuration_info.h
… editor/editor_inspector.cpp
… editor/editor_node.cpp
… editor/export/editor_export_plugin.cpp
… editor/export/editor_export_plugin.h
… editor/export/editor_export_preset.cpp
… editor/export/editor_export_preset.h
… editor/gui/scene_tree_editor.cpp
… editor/gui/scene_tree_editor.h
… editor/icons/NodeWarnings2.svg
… editor/icons/NodeWarnings3.svg
… editor/icons/NodeWarnings4Plus.svg
… editor/plugins/configuration_info_editor_plugin.cpp
… editor/plugins/configuration_info_editor_plugin.h
… editor/register_editor_types.cpp
… editor/scene_tree_dock.cpp
… editor/scene_tree_dock.h
… modules/csg/csg_shape.cpp
… modules/csg/csg_shape.h
… modules/gdscript/gdscript.cpp
… modules/gdscript/gdscript.h
… modules/mono/csharp_script.cpp
… modules/mono/csharp_script.h
… modules/multiplayer/multiplayer_spawner.cpp
… modules/multiplayer/multiplayer_spawner.h