Closes https://github.com/godotengine/godot-proposals/issues/9416
Closes https://github.com/godotengine/godot-proposals/issues/550 again
Supersedes #68420 and #88182

Description coming later, see proposal for more details.

But the summary:

- show configuration information, warnings and errors with different icons
- assign messages to a specific property, show icon in inspector for it
- extend configuration info system to also work for Resources (also supporting sub-inspector)

### Architectural improvements over previous PR

- No compatibility breaking changes. New APIs for everything, old stuff is left untouched but marked deprecated.
- Pulled as much logic out of Node as possible, instead moving it to its own EditorConfigurationInfo class.

### TODO

- [ ] Show severity in the string representation of a config info.
- [ ] Would be nice to have a tooltip when hovering the icon, but since other icons in the Inspector don't have this yet, it can be done in a follow-up PR.
- [ ] Move `configuration_info_changed` signal to some editor class.
- [ ] Properly update Inspector when sub-resources are modified (may need to change the "cache unchanged" logic for this one).
- [ ] Resource messages currently only show in inspector, and only if "property" is specified. Either make "property" required, or also show config info in other places. Propagation of some sort might make sense?
- [ ] Decide for icons for other severities.
- [ ] Find out if removing unused icons is okay.
- [ ] Test the deprecated method calls (config warnings instead of config info calls).
- [ ] I removed some seemingly dead code at `Resource::update_configuration_warning` and `material.cpp`, check that this did not break anything.
- [ ] Update all class signatures to use the new API. Provide a nice helper macro for declaring most important properties of a config info: (message, property?, severity?).
- [ ] Test whether exported games don't break, since some of the config info system is now part of the editor and not in Node anymore.