@RedMser 2024-05-16
> It also makes me wonder whether all `_get_configuration_warnings()` methods (or at least the method bodies) should be placed behind `#ifdef TOOLS_ENABLED` to reduce binary size.

I tried this change (ifdef the entirety of `get_configuration_info`) and it seems to reduce the binary size by 100KB. Doing so for the body only would be a lot more manual labor, not sure if it's worth the hassle.