@RedMser 2024-08-30
This PR should be ready now.

A larger UI change has been made by introducing a configuration info summary in the inspector:

![image](https://github.com/user-attachments/assets/a6e5a5f9-0e0f-4628-b568-dc17a1e928e2)

I'd like feedback on the [proposal](https://github.com/godotengine/godot-proposals/issues/9416) whether this list should show for Resource only, or also for Node like in the current implementation.

I've also rewritten the description, to remove finished TODOs and clarify what this PR exactly does.