@Mickeon 2024-08-30
Quoting [my own comment on the matter](https://github.com/godotengine/godot/issues/88379#issuecomment-1948930761) last time just for the sake for others to remember it, haven't looked at the implementation yet.