@Calinou 2024-08-31
> A larger UI change has been made by introducing a configuration info summary in the inspector:

Looks good! I'd make the text besides the warning icon yellow (`warning_color`) and red text besides the error icon (`error_color`). This helps the warnings/errors stand out more.