@RedMser 2024-08-31
Great idea, here's how it looks after said changes:

![image](https://github.com/user-attachments/assets/64cce5da-e7e0-4d33-ae9e-aa6b24038882)