@KoBeWi 2024-09-21
IMO the inspector warnings should be folded by default. Also why are they clickable?

https://github.com/user-attachments/assets/a35283b5-f125-449a-83b7-d7c97ccd8ade