@RedMser 2024-09-21
> IMO the inspector warnings should be folded by default.

Sure I can change that.

> Also why are they clickable?

Yeah I have a TODO to replace them with RichTextLabel, so they wrap. Had ideas like copy text to clipboard etc, but didn't implement. Itemlist isn't good for this purpose. Will fix soon.

> Using unsupported severity results in broken icon:

Maybe it should instead fallback to "warning" severity and print an error?