@KoBeWi 2024-09-21
Went through the changes and overall looks fine. Though I'm not sure if there is need for supporting Strings in info arrays. It's not needed for compatibility, only for convenience.