@RedMser 2024-09-21
Thanks for the review KoBeWi, I did a first pass which includes all comments marked as resolved. It's a new commit so it's easier to go over just these changes. Going to do the rest in the next few days, since it's larger changes.

> I'm not sure if there is need for supporting Strings in info arrays. It's not needed for compatibility, only for convenience.

It helps ease the transition from the old to the new system (user just has to rename the method), but indeed it wouldn't strictly be needed and currently comes with some extra complexity on the editor. I don't mind keeping it this way, or restricting the API to `Array[Dictionary]`.