@RedMser 2024-09-22
OK new changes, as new commits to help make review easier.

- Replaced ItemList with RichTextLabel for the UI.
- Introduced `ConfigurationInfo` struct which is used everywhere internally, while the scripting API still uses Dictionary/String as before. This made the code a lot nicer ^^
- I changed `get_configuration_info` (the built-in one, not the one exposed to scripting) to use `ConfigurationInfo` struct, and also went and renamed `warnings` -> `infos` inside.
- Some validation was added, like printing error if message is empty, severity is unknown, property does not exist. How to make them not spam? Using `_ONCE` prints would work, but then the "property missing" error would only show once, even if you rename to another still-not-existing property...

> IMO the inspector warnings should be folded by default.

Each time you edit a property, inspector plugins are recreated, so it collapses automatically all the time. But the opposite (auto-expanding on each edit) is also not great.
Where should I store the collapse state so that it can be remembered? I should maybe push the same fix to the `ControlPositioningWarning` where I took most of the code from 👀