@RedMser 2024-10-01
I've rebased to latest master and verified that everything still works. No problems when using the `template_release` either, `update_configuration_(info|warnings)` calls are a no-op as is intended. Not sure if there's anything left to do from my side now.

Once review is done, I will squash the commits to just have a "add the feature" commit and another "update all nodes to use the new APIs" commit. Is this okay or do you prefer one commit with everything squashed together?