@KoBeWi 2024-10-05
> Maybe it should instead fallback to "warning" severity and print an error?

It could fallback to warning, but not sure about the error. It makes sense, but it's being spammed a lot.

https://github.com/user-attachments/assets/448a822e-9eea-4089-bc5f-6d50a45b8359