@RedMser 2024-10-05
> It could fallback to warning, but not sure about the error. It makes sense, but it's being spammed a lot.

I changed that one to `ERR_PRINT_ONCE` but not sure about the ones in `ConfigurationInfo::ensure_valid`. I thought of changing the return type to an error enum value so that the errors could be returned as new configuration infos but didn't think about it much yet.