@KoBeWi 2024-10-05
The problem with ERR_PRINT_ONCE is that it will appear only once per editor session 🤔
Seems like the property error is also spamming. Maybe a way to fix both is adding a deferred error print that can happen at most once per frame. Something like `queue_print_error`, with check for unique messages.