@RedMser 2024-10-05
> adding a deferred error print that can happen at most once per frame. Something like `queue_print_error`, with check for unique messages.

I could try implementing this (maybe as a new PR so this one doesn't get even larger scope?). Couldn't this be implemented the same way as ERR_PRINT_ONCE, but by having a static variable that keeps track of the frame number that the error was printed on, and compare if it's different?