@KoBeWi 2024-10-06
I'd make this "system" local to ConfigurationInfo. You just need a HashSet for errors and a method that prints all of them. Then once an error is to appear, add it to the HashSet instead and call deferred the method if HashSet was empty.