@RedMser 2024-10-06
I implemented it like you suggested. It ""only"" prints 3 errors per update now, which I guess is better than before 😋