@RedMser 2025-04-18
> I also don't know how to reliably profile Godot for regressions in memory or speed

I've created a simple test project that allows instantiating and optionally freeing N objects (tried `Object` and `CanvasGroup`, the latter being deeper in the inheritance tree and having configuration info defined), measuring the time difference (and using task manager to measure memory impact).

With N = 100 000, I could not measure any difference in both categories, between the latest master and this PR, testing both editor and export templates on Windows.

<details>
<summary>Measurements</summary>

|                   | new+free Object (Time) | new Object (Size) | new+free CanvasGroup (Time) | new CanvasGroup (Size) |
| ----------------- | ---------------------- | ----------------- | --------------------------- | ---------------------- |
| Master (editor)   | 40ms                   | 50MiB             | 150ms                       | 237MiB                 |
| PR (editor)       | 40ms                   | 50MiB             | 150ms                       | 237MiB                 |
| Master (template) | 30ms                   | 50MiB             | 140ms                       | 249MiB                 |
| PR (template)     | 30ms                   | 50MiB             | 140ms                       | 248MiB                 |

Note that the values are very noisy, so they might be off by a little bit. But no outliers are visible.

</details>

I've included the project here, if anyone wants to experiment with it: [performance-objects.zip](https://github.com/user-attachments/files/19814574/performance-objects.zip)

I had an issue where the Windows export template executable from the CI/CD uses more idle CPU than my self-compiled one, but not sure what to make of it, since I can't profile the former without debug symbols.